added logic to makesymlinks.sh to only attempt commands based on existing files or symlinks

+++ makesymlinks.sh
@@ -1,24 +1,31 @@
 #!/bin/bash
 #
 # makesymlinks.sh - Jesse Tavizon - iktome@adohi.net
+#
 # symlinks for .[file|directory] in ~/ to matching git sources in ~/dotfiles
+#  - creates ~/dotfiles_old as a backup directory
+#  - creates backup of existing dotfile for each source dotfile
+#  - creates symbolic link for each source dotfile in ~/
 
-# Variables
-dir=~/dotfiles                   # dotfiles directory
-olddir=$dir/dotfiles_old            # original dotfiles backup directory
-files="vimrc"                    # list of files/folders to symlink
+dir=~/dotfiles                   # dotfiles repository directory
+olddir=~/dotfiles_old            # original dotfiles backup directory
+files="vimrc"                    # list of files/folders for symlinks
 
-echo -n "Changing to the $dir directory"
+echo "Changing to the $dir directory"
 cd $dir
-echo "done"
 
-echo -n "Creating $olddir for backup of any existing dotfiles in ~/"
-mkdir -p $olddir
-echo "done"
+if ! [ -d "$olddir" ]; then
+   echo "Creating $olddir for backup of existing dotfiles"
+   mkdir -p $olddir
+fi
 
-for file in $files; do
-    echo "Backing up existing .$file in ~/ if needed"
-    mv ~/.$file $olddir 
-    echo "Creating symlink to $file."
-    ln -s $dir/$file ~/.$file
+for file in $files; do    
+   if [ -f ~/.$file ] && ! [ -L ~/.$file ]; then
+      echo "Backing up existing .$file to $olddir"
+      mv ~/.$file $olddir 
+   fi
+   if ! [ -f ~/.$file ]; then
+      echo "Creating symlink to $file"
+      ln -s $dir/$file ~/.$file
+   fi
 done
